Test: match trio/anyio/asyncio substitution at word boundaries
Previously, the test framework's `replace_library` rewrote every substring
occurrence of "trio" / "anyio" / "asyncio", which mangled unrelated package
names like trio_websocket, trio_util, qtrio, and anyio.from_thread in eval
files. Now identifier-like replacements match at word boundaries; library
names additionally allow a leading `_` so the "_trio" suffix in error
codes like "ASYNC103_trio" still rewrites per library.

With the substitution no longer mangling trio_websocket et al., the
library-specific test file async101_trio_pkgs.py is no longer needed --
those cases merge into async101_third_party.py and are now verified
under all three libraries.
