+++ tests/eval_files/async101_third_party.py
@@ -1,7 +1,9 @@
-# 3rd-party context managers with internal cancel scopes / nurseries.
-# These names don't contain "trio" / "anyio" as substrings (or only contain
-# them in a position that survives the test framework's library substitution),
-# so this file is checked under all three libraries.
+# 3rd-party context managers with internal cancel scopes / nurseries /
+# task groups. `yield`ing inside any of them breaks exception handling.
+#
+# Package names like `trio_websocket` / `qtrio` survive the test framework's
+# `trio` -> `anyio` / `asyncio` library substitution because that substitution
+# only matches at word boundaries.
 from contextlib import asynccontextmanager
 
 import anyio.from_thread
@@ -9,6 +11,69 @@ import apscheduler
 import asgi_lifespan
 import mcp.client.sse
 import mcp.client.streamable_http
+import qtrio
+import trio_asyncio
+import trio_parallel
+import trio_util
+import trio_websocket
+
+
+# trio_websocket
+async def foo_open_websocket():
+    async with trio_websocket.open_websocket("h", 80, "/", use_ssl=False) as _:
+        yield 1  # error: 8
+
+
+async def foo_open_websocket_url():
+    async with trio_websocket.open_websocket_url("ws://x") as _:
+        yield 1  # error: 8
+
+
+async def foo_serve_websocket():
+    async with trio_websocket.serve_websocket(
+        lambda *_: None, "h", 80, ssl_context=None
+    ) as _:
+        yield 1  # error: 8
+
+
+@asynccontextmanager
+async def foo_trio_websocket_safe():
+    async with trio_websocket.open_websocket_url("ws://x") as _:
+        yield 1  # safe
+
+
+# trio_asyncio
+async def foo_open_loop():
+    async with trio_asyncio.open_loop() as _:
+        yield 1  # error: 8
+
+
+# trio_parallel
+async def foo_open_worker_context():
+    async with trio_parallel.open_worker_context() as _:
+        yield 1  # error: 8
+
+
+# trio_util
+async def foo_move_on_when():
+    async with trio_util.move_on_when(lambda: None) as _:
+        yield 1  # error: 8
+
+
+async def foo_run_and_cancelling():
+    async with trio_util.run_and_cancelling(lambda: None) as _:
+        yield 1  # error: 8
+
+
+# qtrio
+async def foo_open_emissions_nursery():
+    async with qtrio.open_emissions_nursery() as _:
+        yield 1  # error: 8
+
+
+async def foo_enter_emissions_channel():
+    async with qtrio.enter_emissions_channel(signals=()) as _:
+        yield 1  # error: 8
 
 
 # asgi_lifespan
